ml_worker/requirements.txt
androguard==3.3.5
scikit-learn
joblib
numpy
google-generativeai